# pkg/docs/swagger.yaml
basePath: /api/v0
...
host: localhost:8081
info:
  contact:
    email: info@skypier.io
    name: Skypier
    url: http://skypier.io/
  description: Skypier - Embark securely on web3
  license:
    name: MIT
    url: https://choosealicense.com/licenses/mit/
  title: Skypier
  version: 0.0.1
paths:
...
  /connect/{peerId}:
    get:
      description: Connect to a remote libp2p peer and enable new Stream
      parameters:
      - description: Peer ID
        in: path
        name: peerId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Connect to a remote libp2p peer and enable new Stream
      tags:
      - VPN
...
swagger: "2.0"